fix: Add test to assure `echo "$@" >&2` works as expected (#1284)
Previously it would print the arguments twice, now it only prints
it once as we avoid adding the 'trick' to the script if it's already
present.

This helps with custom git credential helpers who endorse to
to call them like `!credential-helper $@`, in which case the
added arguments are already part of the script.

See https://github.com/languitar/pass-git-helper for detailed
documentation on how that should usually work.

Git seems to manage to not automatically add '$@' when calling
credential helpers, but it's something that is done by `gix-command`
automatically when a command should be invoked that receives arguments
*and* has to be evaluated by a shell.

The current implementation is very naive, but should also work
for 99.9% of the cases out there.